// runtime/save-data/types.ts
export type AccountSeed = {
  uid: string;
  username: string;
  nickname: string;
};

export type Account = AccountSeed & {
  id: number;
};